remove binaries for plugins that are now built-ins
