+++ cmd/infrakit/main.go
@@ -41,6 +41,7 @@ import (
 
 	// Supported "kinds"
 	_ "github.com/docker/infrakit/pkg/run/v0/aws"
+	_ "github.com/docker/infrakit/pkg/run/v0/combo"
 	_ "github.com/docker/infrakit/pkg/run/v0/enrollment"
 	_ "github.com/docker/infrakit/pkg/run/v0/file"
 	_ "github.com/docker/infrakit/pkg/run/v0/group"
